+++ admin/view/javascript/nuvei_version_checker.js
@@ -1,4 +1,4 @@
 window.onload = (event) => {
     $('#content .container-fluid:first')
-        .append('<div class="alert alert-info"><i class="fa fa-info-circle"></i> There is <a href="https://github.com/Nuvei/nuvei-plugin-opencart-3/blob/main/CHANGELOG.md" target="_blank">newer version</a> for Nuvei Checkout plugin.</div>');
+        .append(`<div class="alert alert-info"><i class="fa fa-info-circle"></i> ${nuveiPluginUpgradeNotification}</div>`);
 };